+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 * [BREAKING] Replaced `messages` declaration with separate `success` and `error` calls
 * [BREAKING] Removed `rescues` method (use `error_from` for custom error messages; all exceptions now report to `on_exception` handlers)
 * [BREAKING] Replaced `error_from` with an optional `if:` argument on `error`
-* [FEAT] Implemented conditional success message filtering as well
+  * [FEAT] Implemented conditional success message filtering as well
+* [FEAT] Added block support for `error` and `success`
 
 ## 0.1.0-alpha.2.6.1
 * [FEAT] Added `elapsed_time` and `outcome` methods to `Action::Result`